backend/requirements.txt
# ZyrexAi Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database & ORM
sqlmodel==0.0.14
aiosqlite==0.19.0

# Vector Database & Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1
numpy<2.0.0

# LLM Integration
httpx
ollama==0.1.6

# Tools
duckduckgo-search==4.1.1
PyPDF2==3.0.1
python-docx==1.1.0
psutil==5.9.6

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Scheduling
apscheduler==3.10.4
croniter==2.0.1

# Logging
loguru==0.7.2

# CORS & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
ruff==0.1.11
